better tokenization: uses new command struct, for ease of interpreting complex input
this is in preperation for implementing the pipeline operator, which requires child processes to each handle a seperate command in the input stream. now, rather than having to share sections of one big string array, each one only needs to get a 'command' structure.

this will also simplify future implementations of other operators, such as: >, >>, &, ||...